+++ user-service/src/main/java/com/paypal/user_service/service/UserService.java
@@ -10,7 +10,7 @@ public interface UserService {
 
     User createUser(User user);
 
-    Optional<User> getUserById(int id);
+    Optional<User> getUserById(Long id);
 
     List<User> getAllUsers();
 }
